Table updated for database

+++ src/application/DatabaseTable.java
@@ -2,10 +2,10 @@ package application;
 
 public class DatabaseTable {
 	private String month;
-	private int wattage, gallons, cost;
+	private double wattage, gallons, cost;
 	
 	// Standard object creator for DatabaseTable
-	public DatabaseTable(String month, int wattage, int gallons, int cost) {
+	public DatabaseTable(String month, Double wattage, Double gallons, Double cost) {
 		this.month = new String(month);
 		this.wattage = wattage;
 		this.gallons = gallons;
@@ -21,7 +21,7 @@ public class DatabaseTable {
 	}
 	
 	// Getter and setter for the wattage column data
-	public int getWattage() {
+	public double getWattage() {
 		return wattage;
 	}
 	public void setWattage(int wattage) {
@@ -29,7 +29,7 @@ public class DatabaseTable {
 	}
 	
 	// Getter and setter for the gallons column data
-	public int getGallons() {
+	public double getGallons() {
 		return gallons;
 	}
 	public void setGallons(int gallons) {
@@ -37,7 +37,7 @@ public class DatabaseTable {
 	}
 	
 	// Getter and setter for the cost column data
-	public int getCost() {
+	public double getCost() {
 		return cost;
 	}
 	public void setCost(int cost) {
